Allow uploading UploadedFiles which aren't remote
Not every UploadedFile needs to have an HTTP/FTP URL, it can be stored
in a database, or it can be from FileSystem where the user didn't set a
`:host` so the URL is relative. If this happens, we can download the
file, and there generally shouldn't be any performance penalty, because
a lot of storages will probably alias #open to #download (#open is used
internally if UploadedFile is passed directly).

We need to do this because rest-client currently doesn't accept non-file
IOs. However, there is an open PR for that:

https://github.com/rest-client/rest-client/pull/449