Fix composing generating id of uploaded file
The edge case that wasn't covered is that, if you upload a file in "raw"
mode to a public_id without extension, Cloudinary will still add an
extension it determines, so we still need to update the id with that
extension.

Testing with Shrine's latest Linter revealed this error, because it
stopped using extensions on ids. This made the test with "raw" mode
fail, so we have this tested.