Make storage able to upload arbitrary IOs
Rest-client has a limitation that, in order to determine whether we're
uploading a file, it checks if the object responds to #read and #path.
This obviously doesn't work in Shrine, because not all IO objects are
files. I opened a PR for this some time ago, but so far it's just
sitting there:

https://github.com/rest-client/rest-client/pull/449

We've currently gotten around that by downloading the file in case it's
a Shrine::UploadedFile. However, this still wouldn't work for other IO
objects, e.g. StringIO.

So, in order to pass rest-client's check, we forcefully define #path if
it doesn't already exist. We also override #original_filename, because
in case it's blank the upload will fail with a "missing parameter --
file" error. Rest-client still adds `filename=""` in the multipart
request, so my guess is that because of this the remote server doesn't
recognize the file.